Add files via upload
This the python code to solve numerically the OSP influence problem, using the two methods: matrix inversion and iterative solution. The specific program and algorithm are suggested for small sizes of the social graph, say less than 1K users. A sparse version of the algorithm that can solve the influence problem for millions of users is included in a separate file.